[Xcode] Track "Additional Resources" copied from WebKitAdditions https://bugs.webkit.org/show_bug.cgi?id=261240 rdar://115082280
Reviewed by Alexey Proskuryakov.

Use a pair of xcfilelists provided by WebKitAdditions to copy resources,
similarly to how WebKit reads other xcfilelists from WKA. This makes the
build system able to track when these files are changed/removed, and
rerun the script accordingly. It also works towards making WebKit's
module verifier task not re-run spuriously every build.

Generalize the "migrate headers" script used to copy headers respecting
exclude rules, renaming it "xcfilelist-copy.py", and teach it to
optionally perform a plain copy instead of invoking a build rule script.

* Source/WebKit/Configurations/Base.xcconfig:
* Source/WebKit/Configurations/DebugRelease.xcconfig:
* Source/WebKit/MigratedHeaders-output.xcfilelist:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Source/WebKitLegacy/MigratedHeaders-output.xcfilelist:
* Source/WebKitLegacy/WebKitLegacy.xcodeproj/project.pbxproj:
* Source/WebKitLegacy/mac/Configurations/WebKitLegacy.xcconfig:
* Source/WebKitLegacy/scripts/xcfilelist-copy.py: Renamed from Source/WebKitLegacy/scripts/migrate-headers.py.
(pattern_match):

Canonical link: https://commits.webkit.org/267747@main
