+++ Source/WebKit/Configurations/DebugRelease.xcconfig
@@ -56,10 +56,12 @@ WK_DEFAULT_LTO_MODE = $(WK_USER_LTO_MODE_none);
 WK_CCACHE_DIR = $(SRCROOT)/../../Tools/ccache;
 #include "../../../Tools/ccache/ccache.xcconfig"
 
-WK_WEBKIT_ADDITIONS_PATH = ../../../Internal/WebKit/WebKitAdditions/Additions/WebKit
-WK_WEBKIT_DERIVEDSOURCES_INPUT_XCFILELIST_ADDITIONS = $(WK_WEBKIT_DERIVEDSOURCES_INPUT_XCFILELIST_ADDITIONS_$(USE_INTERNAL_SDK))
-WK_WEBKIT_DERIVEDSOURCES_INPUT_XCFILELIST_ADDITIONS_ =
-WK_WEBKIT_DERIVEDSOURCES_INPUT_XCFILELIST_ADDITIONS_YES = $(WK_WEBKIT_ADDITIONS_PATH)/DerivedSources-input.xcfilelist
-WK_WEBKIT_DERIVEDSOURCES_OUTPUT_XCFILELIST_ADDITIONS = $(WK_WEBKIT_DERIVEDSOURCES_OUTPUT_XCFILELIST_ADDITIONS_$(USE_INTERNAL_SDK))
-WK_WEBKIT_DERIVEDSOURCES_OUTPUT_XCFILELIST_ADDITIONS_ =
-WK_WEBKIT_DERIVEDSOURCES_OUTPUT_XCFILELIST_ADDITIONS_YES = $(WK_WEBKIT_ADDITIONS_PATH)/DerivedSources-output.xcfilelist
+WK_WEBKIT_ADDITIONS_PATH = ../../../Internal/WebKit/WebKitAdditions/Additions/WebKit;
+WK_WEBKIT_DERIVEDSOURCES_INPUT_XCFILELIST_ADDITIONS = $(WK_WEBKIT_DERIVEDSOURCES_INPUT_XCFILELIST_ADDITIONS_$(USE_INTERNAL_SDK));
+WK_WEBKIT_DERIVEDSOURCES_INPUT_XCFILELIST_ADDITIONS_YES = $(WK_WEBKIT_ADDITIONS_PATH)/DerivedSources-input.xcfilelist;
+WK_WEBKIT_DERIVEDSOURCES_OUTPUT_XCFILELIST_ADDITIONS = $(WK_WEBKIT_DERIVEDSOURCES_OUTPUT_XCFILELIST_ADDITIONS_$(USE_INTERNAL_SDK));
+WK_WEBKIT_DERIVEDSOURCES_OUTPUT_XCFILELIST_ADDITIONS_YES = $(WK_WEBKIT_ADDITIONS_PATH)/DerivedSources-output.xcfilelist;
+WK_WEBKIT_RESOURCES_INPUT_XCFILELIST_ADDITIONS = $(WK_WEBKIT_RESOURCES_INPUT_XCFILELIST_ADDITIONS_$(USE_INTERNAL_SDK));
+WK_WEBKIT_RESOURCES_INPUT_XCFILELIST_ADDITIONS_YES = $(WK_WEBKIT_ADDITIONS_PATH)/AdditionalResources-input.xcfilelist
+WK_WEBKIT_RESOURCES_OUTPUT_XCFILELIST_ADDITIONS = $(WK_WEBKIT_RESOURCES_OUTPUT_XCFILELIST_ADDITIONS_$(USE_INTERNAL_SDK));
+WK_WEBKIT_RESOURCES_OUTPUT_XCFILELIST_ADDITIONS_YES = $(WK_WEBKIT_ADDITIONS_PATH)/AdditionalResources-output.xcfilelist
