+++ Source/WebKitLegacy/mac/Configurations/WebKitLegacy.xcconfig
@@ -122,7 +122,7 @@ OTHER_TAPI_FLAGS = $(inherited) -fvisibility-inlines-hidden $(WK_CFLAGS_BUILD_FO
 OTHER_TAPI_FLAGS_SRCROOT_YES = -Xparser -include -Xparser mac/WebKitPrefix.h $(OTHER_TAPI_FLAGS_102197642_WORKAROUND);
 
 // This unfortunate mess of arguments mirrors the EXCLUDED_ and INCLUDED_SOURCE_FILE_NAMES logic above for migrated headers. It needs to be kept in sync to avoid TAPI errors.
-// FIXME: Convert to a filelist generated by migrate-headers.py once rdar://102197642 is completed.
+// FIXME: Convert to a filelist generated by xcfilelist-copy.py once rdar://102197642 is completed.
 OTHER_TAPI_FLAGS_102197642_WORKAROUND = $(OTHER_TAPI_FLAGS_102197642_WORKAROUND_$(WK_COCOA_TOUCH)) -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WebKitAvailability.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WebScriptObject.h;
 OTHER_TAPI_FLAGS_102197642_WORKAROUND_cocoatouch = $(OTHER_TAPI_FLAGS_102197642_WORKAROUND_cocoatouch_$(USE_INTERNAL_SDK)) $(OTHER_TAPI_FLAGS_102197642_WORKAROUND_$(ENABLE_IOS_TOUCH_EVENTS)) -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/AbstractPasteboard.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/KeyEventCodesIOS.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WAKAppKitStubs.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WAKResponder.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WAKView.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WAKWindow.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WebCoreThread.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WebCoreThreadMessage.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WebCoreThreadRun.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WebEvent.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WebItemProviderPasteboard.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WKContentObservation.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WKGraphics.h -extra-private-header $(TARGET_BUILD_DIR)/$(PRIVATE_HEADERS_FOLDER_PATH)/WKTypes.h;
 OTHER_TAPI_FLAGS_102197642_WORKAROUND_cocoatouch_YES = $(OTHER_TAPI_FLAGS_102197642_WORKAROUND_ENABLE_IOS_TOUCH_EVENTS);
